requirements.txt
litserve
torch
torchvision
pillow
numpy
typing
scikit-image
huggingface_hub
transformers>=4.39.1
